requirements.txt
flask
pdfplumber
pytesseract
pillow
transformers
flask_cors
